fix: "인증 미들웨어 user 객체 사용례 수정"
req를 user 객체로 할당하도록 변경

+++ src/routes/char.router.js
@@ -12,19 +12,19 @@ const router = express.Router();
  */
 router.post("/char", authMiddleware, async (req, res, next) => {
   const { name } = req.body;
-  const { user } = req.user;
+  const { user } = req;
 
   try {
     // 캐릭터명 중복 여부
     const isExistName = await prisma.characters.findFirst({
-      where: { name: name, user_id: user },
+      where: { name: name, user_id: +user.user_id },
     });
     if (isExistName) throw throwError("이미 존재하는 캐릭터명입니다.", 409);
 
     const char = await prisma.characters.create({
       data: {
         name: name,
-        user_id: +user,
+        user_id: +user.user_id,
       },
     });
 
@@ -41,12 +41,12 @@ router.post("/char", authMiddleware, async (req, res, next) => {
  */
 router.get("/char/:char_id", optionalAuthMiddleware, async (req, res, next) => {
   const { char_id } = req.params;
-  const user = req.user;
+  const { user } = req;
 
   try {
     // 캐릭터 존재 여부
     const char = await prisma.characters.findFirst({
-      where: { char_id: +char_id },
+      where: { char_id: +char_id }, // user_id 제외
     });
     if (!char) throw throwError("캐릭터가 존재하지 않습니다.", 404);
 
@@ -73,15 +73,16 @@ router.get("/char/:char_id", optionalAuthMiddleware, async (req, res, next) => {
  */
 router.delete("/char/:char_id", authMiddleware, async (req, res, next) => {
   const { char_id } = req.params;
-  const { user } = req.user;
+  const { user } = req;
 
   try {
     // 캐릭터 존재 여부
-    await checkChar(prisma, char_id, user);
+    const char = await checkChar(prisma, char_id, user.user_id);
+    if (!char) throw throwError("캐릭터 정보가 없거나 권한이 없습니다.", 403);
 
     // 캐릭터 삭제
     const deletedChar = await prisma.characters.delete({
-      where: { char_id: +char_id, user_id: user },
+      where: { char_id: +char_id, user_id: user.user_id },
     });
 
     return res
@@ -100,11 +101,11 @@ router.get(
   authMiddleware,
   async (req, res, next) => {
     const { char_id } = req.params;
-    const { user } = req.user;
+    const { user } = req;
 
     try {
       // 캐릭터 존재 여부
-      await checkChar(prisma, char_id, user);
+      await checkChar(prisma, char_id, user.user_id);
 
       // 인벤토리의 아이템 목록 조회
       const inventory = await prisma.character_inventory.findMany({
@@ -132,4 +133,41 @@ router.get(
     }
   }
 );
+
+/**
+ * 장착한 아이템 목록 조회 API
+ */
+router.get("/char/equip/:char_id", async (req, res, next) => {
+  const { char_id } = req.params;
+
+  try {
+    // 캐릭터 존재 여부
+    await prisma.characters.findFirst({
+      where: { char_id: +char_id },
+    });
+
+    // 장착한 아이템 목록 조회
+    const equip = await prisma.character_item.findMany({
+      where: { char_id: +char_id },
+      select: {
+        item_code: true,
+        item: {
+          select: {
+            item_name: true,
+          },
+        },
+      },
+    });
+
+    // 장착한 아이템 데이터 가공
+    const result = equip.map((x) => ({
+      item_code: x.item_code,
+      item_name: x.item.item_name,
+    }));
+
+    return res.status(200).json({ data: result });
+  } catch (error) {
+    next(error);
+  }
+});
 export default router;
